test: remove() 함수 테스트하기

+++ src/user/user.service.ts
@@ -20,8 +20,8 @@ export class UserService {
     return this.userRepository.find();
   }
 
-  findOne(id: number) {
-    const user = this.userRepository.findOne({ where: { id } });
+  async findOne(id: number) {
+    const user = await this.userRepository.findOne({ where: { id } });
     if (!user) {
       throw new NotFoundException('존재하지 않는 사용자입니다.');
     }
@@ -37,7 +37,13 @@ export class UserService {
     return await this.userRepository.findOne({ where: { id } });
   }
 
-  remove(id: number) {
-    return this.userRepository.delete(id);
+  async remove(id: number) {
+    const user = await this.userRepository.findOne({ where: { id } });
+    if (!user) {
+      throw new NotFoundException('존재하지 않는 사용자입니다.');
+    }
+
+    await this.userRepository.delete(id);
+    return id;
   }
 }
